toolchain fixes
A toolchain is only a toolchain *file*

Everything else (the name of the toolchain to build, whether we
are cross-compiling or not, whether missing packages are to important or
not), comes from here.

Get rid of the --toolchain option, use --toolchain-file instead.

The only settings that matters now is
general.build.toolchain_file

general.build.toolchain is no longer used.

Add a nice `qitoolchain use' action