cmake: post-copy-dlls: much simpler code
We chose to every DLL we can think of.
This will make the first build much longer, but we are sure to never forget one
dll, and it's only done once.

So we can get rid of the post-build rule, and simply copy the dlls
ant the end of qibuild/general

FIXME: update doc!