cmake: add support for SYSCONFDIR override
By default the config location is set to 'etc'.
This is wrong when an install prefix is applie; in such cases, things got
installed in:
  ${CMAKE_INSTALL_PREFIX}/etc/

The former install layout was (with CMAKE_INSTALL_PREFIX=/usr):
/
└── usr
    ├── etc
    │   └── foo
    │       └── foo.cfg
    └── lib
        └── libfoo.so

This patch adds to possobility to override the sysconfdir location
(commonly used on UNIX system) to an absolute path, which tells CMake to
not apply the CMAKE_INSTALL_PREFIX on the configuration files (i.e. when
calling qi_install_conf).

The resulting layout is (with CMAKE_INSTALL_PREFIX=/usr):
/
├── etc
│   └── foo
│       └── foo.cfg
└── usr
    └── lib
        └── libfoo.so

Change-Id: I27077abac6e55956dd8d7a9558f999f0aef4729b
Reviewed-on: http://gerrit.aldebaran.lan:8080/9658
Reviewed-by: dmerejkowsky <dmerejkowsky@aldebaran-robotics.com>
Reviewed-by: smartin <smartin@aldebaran-robotics.com>
Tested-by: smartin <smartin@aldebaran-robotics.com>