qi_stage_script: Reimplement using an optional trampoline.
Rationale: We don't want to stage all non-generated stuff that
a script might need. So better leave the script in srcdir, put a
config.cmake to find it, and if needed, provide a trampoline
in sdk/bin that bounces to the script in src dir after setting
some environment variables.

Change-Id: I5f7513807a2c7fe8d491f266a7daf3c579647583
Reviewed-on: http://gerrit.aldebaran.lan:8080/18595
Reviewed-by: dmerejkowsky <dmerejkowsky@aldebaran-robotics.com>
Reviewed-by: mnottale <mnottale@aldebaran-robotics.com>
Tested-by: mnottale <mnottale@aldebaran-robotics.com>