qisrc: add support for tags
This commit lets users specify a tag name (or even a commit hash) in the
`branch` attribute of a repo in the manifest.xml file.

This used to fail because qisrc tried to checkout `remote_name`/`branch`, which
is not valid when `branch` is actually a tag.

The proposed solution is entirely back-compatible: it will only alter the
current behavior when `remote_name`/`branch` is invalid anyway (checked with
`git rev-parse`). When this is the case, it will try to checkout `branch`
rather than `remote_name`/`branch`, which will work for tags.