qisrc: handle relative paths
This commit lets users specify a relative path as the `manifest_url` argument
of `qisrc init`.

This used to fail because the actual `git fetch` is performed from
`.qi/manifests/default`, defeating the relative path.

The proposed solution is to convert relative paths to absolute ones before
passing them to configure_manifest. This is only done when the given url is
a valid local path.

Change-Id: I30c128ecd3b59e54bf9abe9490e6bb6b3b6d68eb
Reviewed-on: http://gerrit.aldebaran.lan/60151
Tested-by: gerrit
Reviewed-by: dmerejkowsky <dmerejkowsky@aldebaran-robotics.com>