qitoolchain: add checksum support
Feeds can now specify a checksum attribute for each downloaded package,
and qitoolchain will check it upon download. A warning is printed if the
attribute is not specified, and an error occurs if the attribute is
specified but wrong.

To update the current feeds with checksums, a "--update-checksums"
switch is provided.

For the sake of --update-checksums, this commit makes "<feed name=" work
from local feeds and regular URLs as well as Git.

In case strict_metadata is false and package.xml has the wrong name,
there is some extra handling to make sure we update the right name in
the feed.

Limitations:
 - This does not apply to SVN packages yet.
 - Checksums are checked on downloaded *packages* only; no checksum is
   calculated on downloaded *feeds*.

Change-Id: I2ccc3104d808f2dee6cc52dbb20bdf3e48842b86
Reviewed-on: http://gerrit.aldebaran.lan/73469
Tested-by: gerrit
Reviewed-by: vbarbaresi <vbarbaresi@aldebaran-robotics.com>