qitoolchain: fix breakage introduced by post_add
The post_add method was calling method "load_package_xml", which was
overriding variables written by method "reroot_path".

Remove the call to load_package_xml in qipackage, as it is always called
in database.py before calling post_add.

The API is rather brittle, let's fix it in a subsequent commit

Change-Id: I0f9b4012bc368eb66b882fb50b0f2a0ae3f9d870
Reviewed-on: http://gerrit.aldebaran.lan/78081
Reviewed-by: plemagourou <plemagourou@aldebaran.com>
Tested-by: gerrit