qitoolchain: Add sub package feed parsing.
In the toolchain xml file, a 'package' tag contained in an other
'package' tag is not taken into account. This patch adds a feature that
look for 'package' tags inside 'package' tags.

Now, this kind of xml file is valid:
<toolchain strict_metadata="false">
  <package name="meta-package" ...>
    <package name="sub-pkg1" version="..."/>
    <package name="sub-pkg2" version="..."/>
    <package name="sub-pkg3" version="..."/>
  </package>
</toolchain>

This is needed when a package (eg. yocto-sdk) provides several
sub-packages that need to appear as qitoolchain packages.

The xml sub package tag should not contain 'directory' or 'url'
attributes as 'directory' attribute will be set to the parent package
'directory' attribute.

Change-Id: Ib7b1f22a065c30c9d9616383fbd7629f73b051e2
Reviewed-on: http://gerrit.aldebaran.lan/79917
Reviewed-by: nrubinstein <nrubinstein@aldebaran-robotics.com>
Tested-by: plemagourou <plemagourou@aldebaran.com>