breakpad: do not change permissions when dumping symbols
there is no reason for doing so, it is probably a leftover
from when symbols were systematically stripped when dumped.

With this change it becomes possible to strip symbols from
binaries on a read-only filesystem.

Also change the can_be_dumped() helper, to stop checking whether
the binary file can be written to:

* write access is not necessary to dump symbols from the binary
  (it is only necessary when stripping symbols)

* the check was flawed anyway: instead of checking that the user
  running qibuold can write to the file, it checked that the file
  owner could, which is irrelevant.

Change-Id: I6c8509d92696214c1ef2976f33e083d7d4c478fe