qisrc: accept sub-qiprojects from submodules
Before, qisrc stop with an error seeing that a sub-qiproject (known to
qisys' worktree) already exists with the same path. In the case of
submodules, it is normal, and should not cause an error.
Now it is acceptable for qisrc, that will only care about registering
the git project properly.