qitoolchain: drop subfeed ``path`` attribute support
subfeed could in theory be included using the ``url`` attribute and/or
the ``path`` attribute.

Since using the ``url`` is the superior solution (see below), let drop
support for the ``path`` alternative, which was undocumented anyway.

It makes qitoolchain simple and fixes ambiguous cases were both
attributes were used.

When using the path attribute, the feed could only be used with a git URL.
This is annoying.

When using the url attribute, the feed can be used in many ways:

with a git URL:

  qitoolchain create  --feed-name linux64-atom --branch release-2.9 mytoolchain git@gitlab.lan:qi/toolchains.git

with a remote URL (here https):

  qitoolchain create mytoolchain  https://gitlab.aldebaran.lan/qi/toolchains/-/raw/release-2.9/feeds/linux64-atom.xml

or with a local URL (typically a local clone of the git repo):

  qitoolchain create mytoolchain ~/src/tools/toolchains/feeds/linux64-atom.xml